backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pillow
python-dotenv
google-adk
cryptography
